whisper_requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
filelock==3.16.1
fsspec==2024.12.0
idna==3.10
importlib-metadata==8.5.0
jinja2==3.1.5
llvmlite==0.41.1
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
openai-whisper==20240930
pillow==10.4.0
regex==2024.11.6
requests==2.32.3
sympy==1.13.3
tiktoken==0.7.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
